+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [17.5.0](https://github.com/ogre-works/ogre-tools/compare/v17.3.0...v17.5.0) (2023-06-21)
+
+### Features
+
+- Introduce InjectWithMeta ([24e9bf0](https://github.com/ogre-works/ogre-tools/commit/24e9bf05f995720790daee31ff0623999e5ce227))
+
+## [17.4.0](https://github.com/ogre-works/ogre-tools/compare/v17.3.0...v17.4.0) (2023-06-21)
+
+### Features
+
+- Introduce InjectWithMeta ([24e9bf0](https://github.com/ogre-works/ogre-tools/commit/24e9bf05f995720790daee31ff0623999e5ce227))
+
 ## [17.3.0](https://github.com/ogre-works/ogre-tools/compare/v17.2.0...v17.3.0) (2023-06-19)
 
 ### Features